Model

Model

POMDPTesting.probability_check — Function.

probability_check(pomdp::POMDP) Checks if the transition and observation function of the discrete pomdp have probability mass that sums up to unity for all state-action pairs.

source

POMDPTesting.obs_prob_consistency_check — Function.

obs_prob_consistency_check(pomdp::POMDP)

Checks if the observation function of the discrete pomdp has probability mass that sums up to unity for all state-action pairs.

source

POMDPTesting.trans_prob_consistency_check — Function.

trans_prob_consistency_check(pomdp::Union{MDP, POMDP})

Checks if the transition function of the discrete problem has probability mass that sums up to unity for all state-action pairs.

source