Model

Model

POMDPTesting.has_consistent_distributions — Function.

has_consistent_distributions(m::Union)

Return true if no problems are found in the distributions for a discrete problem. Print information and return false if problems are found.

Tests whether

All probabilities are positive

Probabilities for all distributions sum to 1

All items with positive probability are in the support

source

POMDPTesting.has_consistent_initial_distribution — Function.

has_consistent_initial_distribution(m)

Return true if no problems are found with the initial state distribution for a discrete problem. Print information and return false if problems are found.

See has_consistent_distributions for information on what checks are performed.

source

POMDPTesting.has_consistent_transition_distributions — Function.

has_consistent_transition_distributions(m)

Return true if no problems are found in the transition distributions for a discrete problem. Print information and return false if problems are found.

See has_consistent_distributions for information on what checks are performed.

source

POMDPTesting.has_consistent_observation_distributions — Function.

has_consistent_observation_distributions(m)

Return true if no problems are found in the observation distributions for a discrete POMDP. Print information and return false if problems are found.

See has_consistent_distributions for information on what checks are performed.

source